% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BP_cells}
\alias{BP_cells}
\title{Cell anatomical characteristics for \emph{Betula platyphylla} in Liangshui}
\format{
A \code{data.frame} containing five columns with year(Year),
relative ring width(RRingWidth), lumen area(LA) ,cell wall thickness(CWT)
and cell type(CellType).
}
\source{
A process-based model of xylogenesis and tree-ring formation of
broad-leaved trees (BTR) driven by climate. (not published yet)
}
\usage{
data(BP_cells)
}
\description{
This dataset gives the fiber cell lumen area, fiber cell wall thickness
and vessel cell lumen area for \emph{Betula platyphylla}
at each relative ring width in each year.
}
\keyword{datasets}
