% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BPage}
\alias{BPage}
\title{BTR model age input for \emph{Betula platyphylla}}
\format{
A \code{data.frame} containing 4 columns :

\strong{Year} : is the year of growth,

\strong{age} : is the age of trees corresponding to the year,

\strong{Tage} : is the trend of cambial activaty,

\strong{Lage} : is the trend of cell enlargement.
}
\source{
A process-based model of climate-driven xylogenesis and tree-ring formation in broad-leaved trees (BTR).\link{DOI:10.1093/treephys/tpae127}
}
\usage{
data(BPage)
}
\description{
This dataset gives the BTR model age trend input for simulate
\emph{Betula platyphylla} xylogenesis in Liangshui.
}
\keyword{datasets}
