% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BPparam}
\alias{BPparam}
\title{BTR model parameters for \emph{Betula platyphylla}}
\format{
A \code{data.frame} containing 7 columns :

\strong{Parameter} : is the parameter name in Rcode,

\strong{ParamType} : is the parameter type,

\strong{Module} : is the module running the parameter,

\strong{Note} : is the object for which the parameter is used,

\strong{Description} : is the description of parameter,##'

\strong{Values} : is the parameter value.

\strong{Unit} : is the unit parameter value.
}
\source{
A process-based model of climate-driven xylogenesis and tree-ring formation in broad-leaved trees (BTR).\link{DOI:10.1093/treephys/tpae127}
}
\usage{
data(BPparam)
}
\description{
This dataset gives the BTR model parameters list for simulate
\emph{Betula platyphylla} xylogenesis in Liangshui.
}
\keyword{datasets}
