% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_MainFun.R
\name{btr}
\alias{btr}
\title{Main function calculate tree-ring anatomic}
\usage{
btr(
  clim,
  parameters,
  age,
  syear = NA,
  eyear = NA,
  writeRes = F,
  intraannual = F,
  gTmethod = "Jonhson",
  division = "limit",
  testLim = F,
  CZgR = c(1, 0, 0, 1),
  Pbar = F,
  testMod = F,
  Dcase = "min",
  Named = NULL
)
}
\arguments{
\item{clim}{climate data must have Year，DOY，Temp ，Prec}

\item{parameters}{model parameter date ，we ues a excel data to save it.}

\item{syear}{modeling start year}

\item{eyear}{modeling end year}

\item{writeRes}{auto write res_data, logical value, if 'Ture', output result list as xlsx in file.}

\item{intraannual}{output daily result, logical, if 'Ture', output daily result in file.}

\item{gTmethod}{method of calculate gT ,'VS' and 'Jonhson'}

\item{division}{dealtDivision 'fix' is constant, 'limit' is gM/gV & RCTA control.}

\item{testLim}{use Lage limit mw & ml , Lage in age param_data, regulation cell enlargement by CAmax param .}

\item{CZgR}{growth limit : climate-Cambial Cell, cell wall-gM, cell wall-gV, NewgRtype; 0 for False, 1 for True}

\item{Pbar}{Whether to show progress bar}

\item{testMod}{test Mode, logical, if 'False', Lage&Dage use ratio}

\item{Dcase}{How to calculate Division Limit, including: "min","mean","multiply".}

\item{Cores}{核心数}
}
\value{
a excel data
}
\description{
Main function calculate tree-ring anatomic
}
