% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_SubMainFun.R
\name{btr_lite}
\alias{btr_lite}
\title{Submain function calculate tree-ring anatomic}
\usage{
btr_lite(clim, parameters, syear = NA, eyear = NA, mCRD = 16)
}
\arguments{
\item{clim}{climate data must have Year，DOY，Temp ，Prec}

\item{parameters}{model parameter date ，we ues a excel data to save it.}

\item{syear}{modeling start year}

\item{eyear}{modeling end year}

\item{mCRD}{mean CRD}
}
\value{
a excel data
}
\description{
Submain function calculate tree-ring anatomic
}
