% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_DiviADiffer.R
\name{cell_division}
\alias{cell_division}
\title{cell_division function}
\usage{
cell_division(
  clim.today,
  fixparam.divi,
  fixparam.growth.fiber,
  fixparam.growth.vessel,
  dynparam.growth.t,
  cells,
  vessels
)
}
\arguments{
\item{clim.today}{Today climate data}

\item{fixparam.divi}{Fix parameters}

\item{fixparam.growth.fiber}{Fix parameters of fiber cells}

\item{fixparam.growth.vessel}{Fix parameters}

\item{dynparam.growth.t}{Dynamic parameters}

\item{cells}{cells anatomy}

\item{vessels}{vessels anatomy}
}
\value{
today cells layer number & vessels number
}
\description{
cell_division function
}
