% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_CellGrowth.R
\name{cells_growth}
\alias{cells_growth}
\title{cells_growth function}
\usage{
cells_growth(
  cell = GC,
  CorV = "C",
  clim.today,
  layer.max,
  fixparam.growth.fiber,
  fixparam.growth.vessel,
  dynparam.growth.t
)
}
\arguments{
\item{cell}{one of growthing cell layer}

\item{CorV}{cells or vessels}

\item{clim.today}{Today climate data}

\item{layer.max}{max layer in grwoth area}

\item{fixparam.growth.fiber}{Fix parameters}

\item{fixparam.growth.vessel}{Fix parameters}

\item{dynparam.growth.t}{Dynamic parameters}
}
\value{
today cells anatomy
}
\description{
cells_growth function
}
