% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Clim_param}
\alias{Clim_param}
\title{BTR model parameters for microclim module}
\format{
A \code{data.frame} containing 6 columns :

\strong{parameter} : is the parameter name in Rcode,

\strong{description} : is the description of parameter,

\strong{note} : is the note of parameter,

\strong{paramtype} : is the parameter type,

\strong{modul} : is the module running the parameter,

\strong{values} : is the parameter value.
}
\source{
Modifications of Suz Tolwinski-Ward's monthly time-step code by Nick Graham in 2011.
Ported to R by SETW in 2015. Implementation of CPC Leaky Bucket model as described in
Huang et al., 'Analysis of Model-Calculated Soil Moisture over the United States
(1931-1993) and Applications to Long-Range Temperature Forecasts,' J. Clim. (1995)
}
\usage{
data(Clim_param)
}
\description{
This dataset gives the BTR model parameters list for simulate microclim in Liangshui.
Original code from "Run CPC Leaky Bucket model (submonthly version)"
We have modified some of the parameters to allow the model to estimate data on a daily basis.
}
\keyword{datasets}
