% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Climdata.2}
\alias{Climdata.2}
\title{Meteorological data 2 , easily accessible from meteorological stations , input data for the BTR model from Liangshui,China}
\format{
A \code{data.frame} containing 6 columns with year (Year),

\strong{Year} : Year,

\strong{Month} : Month,

\strong{Day} : Day,

\strong{TEM} : Temperature,

\strong{RH} : Relative humidity,

\strong{PRE} : Precipitation.
}
\source{
A process-based model of climate-driven xylogenesis and tree-ring formation in broad-leaved trees (BTR).\link{DOI:10.1093/treephys/tpae127}
}
\usage{
data(Climdata.2)
}
\description{
Model primary inputs include date, temperature, VPD soil moisture at Liangshui, Yichun, Heilongjiang, China.
}
\keyword{datasets}
