% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_MicroClimNeo.R
\name{Compute_clim}
\alias{Compute_clim}
\title{MicroClimate}
\usage{
Compute_clim(climdata, syear = NA, eyear = NA, lat = NA, rootd = 1000, ...)
}
\arguments{
\item{climdata}{Climsdata}

\item{syear}{Start year}

\item{eyear}{End year}

\item{lat}{Latitude of sample sites}

\item{rootd}{Root zone depth}

\item{parameters}{Parameters}
}
\value{
gE , SoilM soil moisture, Ls, potEv et.al. climate factor
SoilM soil moisture computed via the CPC Leaky Bucket model (in v/v, 12 x Nyrs)
potEv Potential evapotranspiration computed via Thornthwaite's 1947 scheme (in mm).
}
\description{
MicroClimate
}
