% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_MicroClimNeo.R
\name{Compute_dayLength}
\alias{Compute_dayLength}
\title{Compute_daylengthfactor}
\usage{
Compute_dayLength(climdata, lat)
}
\arguments{
\item{climdata}{data}

\item{lat}{Latitude}
}
\value{
day length and standardization day length
}
\description{
Compute_daylengthfactor
}
