% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_MicroClim.R
\name{compute_daylengthfactor}
\alias{compute_daylengthfactor}
\title{Compute_daylengthfactor}
\usage{
compute_daylengthfactor(fixparam.Mclim, dynparam.Mclim)
}
\arguments{
\item{fixparam.Mclim}{Fix parameters}

\item{dynparam.Mclim}{Dynamic parameters}
}
\value{
day length and standardization day length
}
\description{
Compute_daylengthfactor
}
