% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_MicroClim.R
\name{Compute_gE}
\alias{Compute_gE}
\title{Sun radiation factor gE}
\usage{
Compute_gE(climdata, fixparam.Mclim, dynparam.Mclim, syear, eyear)
}
\arguments{
\item{climdata}{climsdata}

\item{fixparam.Mclim}{Fix parameters}

\item{dynparam.Mclim}{Dynamic parameters}

\item{syear}{start year}

\item{eyear}{end year}
}
\value{
Sun radiation factor gE
}
\description{
Sun radiation factor gE
}
