% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_5_GR.R
\name{Compute_gR}
\alias{Compute_gR}
\title{Computer gM & gT & gP}
\usage{
Compute_gR(clim, growth_Param)
}
\arguments{
\item{clim}{climates about Temperature & Soil Moisture}

\item{growth_Param}{parameters of climates model}
}
\value{
gM & gT
}
\description{
Computer gM & gT & gP
}
