% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_MicroClimNeo.R
\name{Compute_P}
\alias{Compute_P}
\title{Computer snow to precipitation}
\usage{
Compute_P(climdata)
}
\arguments{
\item{climdata}{climsdata}
}
\value{
climsdata
}
\description{
Computer snow to precipitation
}
