% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_MicroClim.R
\name{Compute_rootd}
\alias{Compute_rootd}
\title{Active soil depth}
\usage{
Compute_rootd(climdata, fixparam.Mclim, dynparam.Mclim)
}
\arguments{
\item{climdata}{climsdata}

\item{fixparam.Mclim}{Fix parameters}

\item{dynparam.Mclim}{Dynamic parameters}
}
\value{
Active soil depth
}
\description{
Active soil depth
}
