% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_MicroClimNeo.R
\name{Computer_soliM}
\alias{Computer_soliM}
\title{cpc-leaky bucket}
\usage{
Computer_soliM(
  climdata,
  paramSoilM = c(M0 = 0.35, dp = 2, alph = 0.0013, Mmax = 0.6, Mmin = 0.042, mu.th =
    1.09, m.th = 4.1)
)
}
\arguments{
\item{climdata}{Climsdata}

\item{paramSoilM}{The parameters related to estimation of soil moisture, where 'M0' is the initial soil moisture, 'dp' is the rainfall infiltration step, 'alph' is the scalar runoff , 'Mmax' is the maximum soil water holding capacity, 'Mmin' is the minimum soil moisture, 'mu.th' is the subsurface runoff parameter, 'm.th' is the surface runoff parameter.}
}
\value{
gE , SoilM soil moisture, Ls, potEv et.al. climate factor
SoilM soil moisture computed via the CPC Leaky Bucket model (in v/v, 12 x Nyrs)
potEv Potential evapotranspiration computed via Thornthwaite's 1947 scheme (in mm).
}
\description{
The main formula is based on VSLiteR cpc-leaky bucket model (submonthly version) and has been modified to fit the daily value data.
VSLiteR:\link{https://github.com/suztolwinskiward}
}
