% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_DailyGrowth.R
\name{daily_grwoth}
\alias{daily_grwoth}
\title{interFun of Daily growth}
\usage{
daily_grwoth(
  newCell,
  newVessel,
  vesselsNum,
  dailyCells,
  cells,
  vessels,
  clim.today,
  fixparam.growth.fiber,
  fixparam.growth.vessel,
  dynparam.growth.t
)
}
\arguments{
\item{newVessel}{形成层活动结果}

\item{vesselsNum}{形成层活动结果}

\item{dailyCells}{细胞们}

\item{cells}{cells anatomy}

\item{vessels}{vessels anatomy}

\item{clim.today}{Today climate data}

\item{fixparam.growth.fiber}{Fix parameters of fiber cells}

\item{fixparam.growth.vessel}{Fix parameters}

\item{dynparam.growth.t}{dynparam.growth}

\item{newCells}{形成层活动结果}
}
\value{
today cells anatomy
}
\description{
interFun of Daily growth
}
