% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_SummaryRes.R
\name{dev.outputs}
\alias{dev.outputs}
\title{Develop Outputs}
\usage{
dev.outputs(BtrResList = Outputs)
}
\arguments{
\item{Result}{rBTR Outputs list}
}
\value{
DevResult result
}
\description{
Develop Outputs
}
