% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FM_param}
\alias{FM_param}
\title{BTR model parameters for \emph{Fraxinus mandshurica}}
\format{
A \code{data.frame} containing 7 columns :

\strong{in_formals} : is the parameter name in Manuscript,

\strong{parameter} : is the parameter name in Rcode,

\strong{description} : is the description of parameter,

\strong{note} : is the object for which the parameter is used,

\strong{paramtype} : is the parameter type,

\strong{modul} : is the module running the parameter,

\strong{values} : is the parameter value.
}
\source{
A process-based model of xylogenesis and tree-ring formation of
broad-leaved trees (BTR) driven by climate. (not published yet)
}
\usage{
data(FM_param)
}
\description{
This dataset gives the BTR model parameters list for simulate
\emph{Fraxinus mandshurica} xylogenesis in Liangshui.
}
\keyword{datasets}
