% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FM_ring}
\alias{FM_ring}
\title{Tree ring width for \emph{Fraxinus mandshurica} in Liangshui}
\format{
A \code{data.frame} containing four columns with year(Year),
mean tree ring width(RingWidth), Standard Deviation(SD) and sample depth(N).
}
\source{
A process-based model of xylogenesis and tree-ring formation of
broad-leaved trees (BTR) driven by climate. (not published yet)
}
\usage{
data(FM_ring)
}
\description{
This dataset gives the mean tree ring width for \emph{Fraxinus mandshurica}
in each year.
}
\keyword{datasets}
