% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LS_clim}
\alias{LS_clim}
\title{Daily Mean Temperature, Total Precipitation and Relative Humidity for
Liangshui,China}
\format{
A \code{data.frame} containing 7 columns with year (Year),
month (Month),day (Day), Day of year (DOY), precipitation (PRE),
temperature (TEM) and relative humidity (RH).
}
\source{
A process-based model of xylogenesis and tree-ring formation of
broad-leaved trees (BTR) driven by climate. (not published yet)
}
\usage{
data(LS_clim)
}
\description{
This dataset gives the daily mean temperature, total
precipitation and relative humidity at Liangshui, Yichun, Heilongjiang, China.
}
\keyword{datasets}
