% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LS_climdata}
\alias{LS_climdata}
\title{Daily Mean Temperature, soil moisture and VPD for
Liangshui,China}
\format{
A \code{data.frame} containing 6 columns with year (Year),
month (Month),day (Day), soil moisture (soilM),
temperature (TEM) and VPD (VPD).
}
\source{
A process-based model of xylogenesis and tree-ring formation of
broad-leaved trees (BTR) driven by climate. (not published yet)
}
\usage{
data(LS_climdata)
}
\description{
This dataset gives the daily mean temperature, soil Moisture and
VPD at Liangshui, Yichun, Heilongjiang, China.
}
\keyword{datasets}
