% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_SummaryRes.R
\name{mod.test.cells}
\alias{mod.test.cells}
\title{Model test Cells}
\usage{
mod.test.cells(
  devOutputs = BtrResList,
  OBSdata,
  celltype = "vessel",
  CellParam = "LA",
  Years,
  method = "pearson",
  k = 5,
  ...
)
}
\arguments{
\item{devOutputs}{Simulates data}

\item{OBSdata}{Observes data From ROXAS cell sheet}

\item{celltype}{'vessel' or 'fiber'}

\item{CellParam}{cell parameter e.g. "LA","WT","WA"}

\item{method}{cor.test method}

\item{cor.test}{time windows (e.g. c(1960:2020) ).}
}
\value{
test result
}
\description{
Model test Cells
}
