% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_SummaryRes.R
\name{mod.test}
\alias{mod.test}
\title{Model test}
\usage{
mod.test(y_actual, y_predicted, method = "pearson", ...)
}
\arguments{
\item{y_actual}{Observes data}

\item{y_predicted}{Simulates data}

\item{method}{cor.test method}
}
\value{
test result
}
\description{
Model test
}
