% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_SummaryRes.R
\name{nor}
\alias{nor}
\title{normalization data}
\usage{
nor(dat, TPH = F, ROU = F, min1 = F, Zeros = F)
}
\arguments{
\item{dat}{a numeric data.}

\item{TPH}{a logical value, if TPH is TRUE, calculate the data as 0-100, otherwise calculate as 0-1.}

\item{ROU}{a logical value, if TPH is TRUE, data will be maintained to two decimal places.}

\item{min1}{a logical value,if TPH is TRUE, data minimum value is not zero.}

\item{Zeros}{a logical value,if TPH is TRUE, raw data minimum value set to 0.}
}
\value{
normalization data.
}
\description{
normalization data
}
