% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_YearGrowth.R
\name{year_growth}
\alias{year_growth}
\title{yearGrowth}
\usage{
year_growth(
  x,
  microclim,
  testMod,
  testLim,
  intraannual,
  writeRes,
  division,
  Dcase,
  CZgR,
  cells,
  vessels,
  dynparam.growth.0,
  RCTA,
  fixparam.divi,
  fixparam.growth.fiber,
  fixparam.growth.origin,
  fixparam.growth.vessel
)
}
\arguments{
\item{microclim}{climate data}

\item{testMod}{locic testMode}

\item{testLim}{loigc Li}

\item{intraannual}{logic output intraannual cells}

\item{writeRes}{logic write results}

\item{division}{分裂}

\item{Dcase}{忘了1}

\item{CZgR}{忘了2}

\item{cells}{fiber cells t0}

\item{vessels}{vessel t0}

\item{dynparam.growth.0}{dynamic param 0}

\item{RCTA}{标准RCTA}

\item{fixparam.divi}{Fix parameters of cell division}

\item{fixparam.growth.fiber}{Fix parameters of fiber cells}

\item{fixparam.growth.origin}{Fis param}

\item{fixparam.growth.vessel}{Fix parameters}

\item{years}{年 输入}

\item{dynparam.growth.t}{dynmaic param t}
}
\value{
today cells anatomy
}
\description{
yearGrowth
}
