% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_MicroClim.R
\name{Compute_clim}
\alias{Compute_clim}
\title{cpc-leaky bucket}
\usage{
Compute_clim(climdata, parameters, syear = NA, eyear = NA)
}
\arguments{
\item{climdata}{climsdata}

\item{parameters}{parameters}

\item{syear}{start year}

\item{eyear}{end year}
}
\value{
gE , gT ,gM , Ls et.al. climate factor
}
\description{
cpc-leaky bucket
}
