% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_MicroClim.R
\name{Compute_clim}
\alias{Compute_clim}
\title{cpc-leaky bucket}
\usage{
Compute_clim(climdata, parameters, syear = NA, eyear = NA)
}
\arguments{
\item{climdata}{climsdata}

\item{parameters}{parameters}

\item{syear}{start year}

\item{eyear}{end year}
}
\value{
gE , SoilM soil moisture, Ls, potEv et.al. climate factor
SoilM soil moisture computed via the CPC Leaky Bucket model (in v/v, 12 x Nyrs)
potEv Potential evapotranspiration computed via Thornthwaite's 1947 scheme (in mm).
}
\description{
#' \code{leakybucket.submonthly} simulates soil moisture with submonthly time step.
}
\details{
Modifications of Suz Tolwinski-Ward's monthly time-step code by Nick Graham in 2011.
Ported to R by SETW in 2015. Implementation of CPC Leaky Bucket model as described in
Huang et al., 'Analysis of Model-Calculated Soil Moisture over the United States
(1931-1993) and Applications to Long-Range Temperature Forecasts,' J. Clim. (1995)
}
