% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_MicroClim.R
\name{Compute_P}
\alias{Compute_P}
\title{Computer snow to precipitation}
\usage{
Compute_P(climdata, fixparam.Mclim, dynparam.Mclim, rootd)
}
\arguments{
\item{climdata}{climsdata}

\item{fixparam.Mclim}{Fix parameters}

\item{dynparam.Mclim}{Dynamic parameters}

\item{rootd}{Active soil depth}
}
\value{
precipitation
}
\description{
Computer snow to precipitation
}
