% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_MicroClimNeo.R
\name{Compute_rootd}
\alias{Compute_rootd}
\title{Active soil depth}
\usage{
Compute_rootd(climdata, rootd0, paramSoilMelt = c(p1 = NA, p2 = NA), Tm = NA)
}
\arguments{
\item{climdata}{climsdata}

\item{rootd0}{root zone depth}

\item{paramSoilMelt}{Soil melt parameters, Where p1 is Soil thawing coefficient 1, p2 is Soil thawing coefficient 2.}

\item{Tm}{Temperature of soil melt}
}
\value{
Active soil depth
}
\description{
Active soil depth
}
